requirements.txt
discord.py
aiocron
requests